requirements.txt
mysql-connector
telethon
cryptg
